requirements.txt
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
mlflow>=2.8.0
fastapi>=0.104.0
uvicorn>=0.24.0
pyyaml>=6.0
joblib>=1.3.0
prometheus-client>=0.18.0
httpx>=0.25.0
pydantic>=2.4.0
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.12.0
datasets>=2.14.0
